requirements.txt
asgiref==3.8.0
autopep8==1.5.7
certifi==2020.6.20
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==4.1
django-allauth==0.42.0
django-ckeditor==5.9.0
django-cors-headers==4.3.0
django-countries==7.5.1
django-js-asset==2.2.0
django-multiselectfield==0.1.12
django-restframework==0.0.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.6
oauthlib==3.2.2
Pillow==8.3.2
pycodestyle==2.7.0
python3-openid==3.2.0
pytz==2021.1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.9.0
urllib3==2.2.0
whitenoise==6.0.0
asgiref==3.8.1
attrs==23.2.0
autopep8==1.5.7
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==4.1
django-allauth==0.48.0
django-cors-headers==4.3.0
djangorestframework==3.15.1
drf-spectacular==0.27.2
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
oauthlib==3.2.2
Pillow==8.3.2
pycodestyle==2.7.0
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0.1
referencing==0.35.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
